Add files via upload
Files from the local computer, after running `2023-11-05_searches_for_manual_editing.Rmd`. Unique document-sentence IDs and manual corrections detected by the Rmd script. Some errors are left: those that would be better to correct in the ConlluEditor. I will do that after the validation check of this batch, since I hope to find unique sentences IDs this time and be able to use the validation check directly as a list of nodes to look at.